Moving request operation protocol to it's own file
