+++ Source/SwiftCloudant/HTTP/RequestBuilder.swift
@@ -16,63 +16,6 @@
 
 import Foundation
 
-// TODO rename to something that makes a little more swift sense.
-/**
- Designates an operation which provides data to perform a HTTP Request.
- */
-internal protocol HTTPRequestOperation   {
-
-    /**
-     The root of url, e.g. `example.cloudant.com`
-     */
-    var rootURL: URL { get }
-    
-    /**
-     The path of the url e.g. `/exampledb/document1/`
-     */
-    var httpPath: String { get }
-    /**
-     The method to use for the HTTP request e.g. `GET`
-     */
-    var httpMethod: String { get }
-    /**
-     The query items to use for the request
-     */
-    var queryItems: [URLQueryItem] { get }
-    
-    /**
-     The body of the HTTP request or `nil` if there is no data for the request.
-     */
-    var httpRequestBody: Data? { get }
-    
-    /**
-     The content type of the HTTP request payload. This is guranteed to be called
-     if and only if `httpRequestBody` is not `nil`
-     */
-    var httpContentType: String { get }
-    
-    /**
-     Provides the `InterceptableSession` to use when making HTTP requests.
-     */
-    var session: InterceptableSession { get }
-    
-    /**
-     A function that is called when the operation is completed.
-     */
-    func completeOperation()
-    /**
-     A function to process the response from a HTTP request.
-
-     - parameter data: The data returned from the HTTP request or nil if there was an error.
-     - parameter httpInfo: Information about the HTTP response.
-     - parameter error: A type representing an error if one occurred or `nil`
-     */
-    func processResponse(data: Data?, httpInfo: HTTPInfo?, error: Error?);
-
-    var isCancelled: Bool { get }
-
-}
-
 /**
  A class which builds `NSURLRequest` objects from `HTTPRequestOperation` objects.
  */
